requirements.txt
py-cord
ezcord
requests
python-dotenv
pytest
aiosqlite